create addCategory func
