+++ Back-End/routes/admin.route.js
@@ -1,18 +1,8 @@
 const adminRouter = require("express").Router();
 
-// // View all courses
-// adminRouter.get('/admin/courses',courses_a)
-// // Hide courses
-// adminRouter.update('/admin/course/:id',hideCourse_a)
-// // Add new categories
-// adminRouter.add('/admin/category',addCategory_a)
-// // View all students accounts details
-// adminRouter.get('/admin/students_details/:id',students_details)
-// // Disable students accounts
-// adminRouter.update('/admin/disable_student/:id',disable_student)
-// // View all instructors accounts details
-// adminRouter.get('/admin/instructors_details/:id',instructors_details)
-// // Disable instructors accounts
-// adminRouter.update('/admin/disable_instructor/:id',disable_instructor)
+const { addCategory} = require('../controllers/main-controller');
+
+// Add new categories
+adminRouter.post('/category',addCategory) // req.body.name
 
 module.exports = adminRouter;
