+++ Back-End/routes/instructors.route.js
@@ -2,13 +2,13 @@ const instructorsRouter = require("express").Router();
 const { visiblityCourse, addCourse,updateCourse,courseRating } = require("../controllers/main-controller");
 
 // Show a hidden course
-instructorsRouter.put('/visiblity/:id',visiblityCourse)
+instructorsRouter.put('/visiblity/:id',visiblityCourse);
 
 // Add courses
-instructorsRouter.post('/course/',addCourse)
+instructorsRouter.post('/course/',addCourse);
 
 // Update courses
-instructorsRouter.put('/course:id/:i',updateCourse) // id = courses.id , i = courses.price
+instructorsRouter.put('/course:id/:i',updateCourse); // id = courses.id , i = courses.price
 
 // View all courses
 // instructorsRouter.get('/courses/',courses)
@@ -17,7 +17,7 @@ instructorsRouter.put('/course:id/:i',updateCourse) // id = courses.id , i = cou
 // instructorsRouter.get('/registeredCourses/:id',registeredCoursesInstructor)
 
 // View course rating
-instructorsRouter.get('/course_rating/:id',courseRating) // id = courses.id 
+instructorsRouter.get('/course_rating/:id',courseRating); // id = courses.id 
 
 // Add live videos
 // instructorsRouter.get('/',)
